Rename RosterContoller.{cpp,h} to RosterController.{cpp,h}
